ext4: disable zeroout optimization when encryption is enabled
When we are manipulate a sparse file's extent tree, there is an
optimization where we will explicitly zero out some data blocks instead
of making the extent tree overly bushy.  Unfortunately, this can
sometimes lead to the file system's extent tree getting corrupted.

Disabling the optimization works around this bug, but we don't have a
completely understood root cause yet.  So this should be understood as
only a (hopefully!) temporary workaround.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: "Theodore Ts'o" <tytso@google.com>
Change-Id: Ie237434603ad1e36351b3e42954ffecbd5c0f035
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>