usb: gadget: f_gsi: Make the header file a local header
All the structures, variables, macros and functions defined in f_gsi.h
are used only by f_gsi.c file. Hence it is not needed to make the
header file f_gsi.h as a public-facing header. Instead move it to
kernel/drivers/usb/gadget/function/ to make it a local header only.

Change-Id: Iabe7625cf266de4f90bbe5f68627a18c5cdcaaff
Signed-off-by: Azhar Shaikh <azhars@codeaurora.org>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>