usb: host: xhci: Abort suspend when interrupt is pending
During xhci suspend driver disables interrupt, clears HW_ACCESSIBLE and
halts the controller. There is still a possibility of a pending interrupt
before controller is halted. If any event is pending with the controller
and SW trying to halt the controller then it resulting in undefined
behavior and leading to assertion of Host controller error(HCE).

Fix this by adding condition to check for interrupt pending during
controller halt. If any event is pending with the controller
then abort suspend.

Change-Id: I38fa6aeeeeccf130280c3af2927fb7f6693f755f
Signed-off-by: Chandana Kishori Chiluveru <cchiluve@codeaurora.org>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>