mdss: import 'sii6400' driver for MHL/WiHD (HDMI Ext.)
Some LeEco devices bundle a Silicon Image chipset for USB to HDMI
conversion using HDMI. However this features is not supported globally
by all of their devices, the Le Pro3 doesn't support it. We're still
adding it to give a purpose to the usage of its setting (i.e. in
smbcharger code).

Change-Id: I3e1d9ca2dc5eb006044d4a53d3fbcf452eca0392
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>