Kconfig, power: adjust LE_AB framework and fix HVDCP3
After a bit of debugging and comparison with the latest stock Kernel
available from OEM, I've finally managed to get the charger code working
almost perfectly.

The main reason it didn't work was due to the OEM Open Kernel having the
order for LE_AB wrong.

In this change we'll beat those two issues by:

1. Making LE_AB framework optional with a new PD configuration.
2. Adjusting voltage measurements to be independent from LE_AB.
3. Making CYCCG USB driver optional as ZL1 doesn't need to use it.
4. Adding a stub skeleton for X2 the next bring-up.
5. Adjusting other dependencies to remove those target specific chargers.

Change-Id: Idb7a40ad3c57868329be105889454f2b9ac6f5a4
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>