staging: introduce an automated qcacld-2.0 repository importer
Automated with Kbuild, when it runs the script checks if the Makefile
exists and if it doesn't it clones it from the remote.

The remote already enables all the options needed for MSM8996 targets.

Supports api.github.com HEAD checking via OAuth token, password or no
auth at all to compare the local HEAD in the driver folder against the
remote one on Github, pulling the updates when available (HEAD->1) or
when the base Kernel tree sets another branch for the driver
(linage-15.1->lineage-14.1).

Change-Id: I73870f3e2d804599ec3d3b4bde469870efdba348
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>